+++ apps/api-journeys-modern/src/lib/theme/theme.zod.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod'
+
+const ThemeModeSchema = z.enum(['dark', 'light'])
+const ThemeNameSchema = z.enum(['base'])
+
+export { ThemeModeSchema, ThemeNameSchema }
